Adding a thermometer server script in micropython.
This required adding some new deps: I'm not sure what the right way to
do logging is with busybox's init. I didn't want to configure the actual
python script differently, so I used some bash magic to redirect all
output to `logger`. That seems to work nicely, but does not work with
`ash` (the shell that comes with busybox).

This feels good enough for now. In the future, we can investigate
streamlining things more:
    - build a rust binary that's ready to go, instead of installing
      micromicropython
    - figure out a better way of capturing stdout/stderr and get rid of
      bash