Remove `ctrl_interface` config.
As of
https://lists.buildroot.org/pipermail/buildroot/2021-February/303700.html,
buildroot defaults to building a version of wpa_supplicant that doesn't
support this `ctrl_interface` parameter. Before this change, I was
seeing `ifup -a` fail with an error like this:

    Line 1: unknown global field 'ctrl_interface=/var/run/wpa_supplicant'
    Line 1: Invalid configuration line 'ctrl_interface=/var/run/wpa_supplicant'

I think this is a unix socket used by a command line tool that I don't
use, so it seems fine to me to just remove this config line entirely.